% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_classifcation.R
\name{get_wb_classification}
\alias{get_wb_classification}
\title{Get ecological classification data for a specified catchment and geography}
\usage{
get_wb_classification(string, column)
}
\arguments{
\item{string}{NAME OF CLASSFICATION AREA E.G. RIVER TILL, THAMES,}

\item{column}{CLASSIFICATION TYPE E.G. OC (Operation Catchment) | MC (Management Catchment) | RBD (River Basin District). See more detail at https://environment.data.gov.uk/catchment-planning/}
}
\value{
ECOLOGICAL CLASSIFICATION OF AREA
}
\description{
Get ecological classification data for a specified catchment and geography
}
