% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_rnag.R
\name{get_wb_rnag}
\alias{get_wb_rnag}
\title{Get 'reasons for not achieving good' data for a specified catchment and geography}
\usage{
get_wb_rnag(string, column)
}
\arguments{
\item{string}{NAME OF CLASSFICATION AREA E.G. RIVER TILL, THAMES,}

\item{column}{CLASSIFICATION TYPE E.G. OC (Operation Catchment) | MC (Management Catchment) | RBD (River Basin District). See more detail at https://environment.data.gov.uk/catchment-planning/}
}
\value{
dataframe
}
\description{
Get 'reasons for not achieving good' data for a specified catchment and geography
}
