% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_sf.r
\name{get_wb_sf}
\alias{get_wb_sf}
\title{Fetch GeoJson data for waterbody}
\usage{
get_wb_sf(string, column)
}
\arguments{
\item{string}{NAME OF CLASSFICATION AREA E.G. RIVER TILL, THAMES,}

\item{column}{CLASSIFICATION TYPE E.G. OC | MC | RBD}
}
\value{
Spatial dataframe
}
\description{
Fetch GeoJson data for waterbody
}
